Compiled the code and renamed to novel binge, so novel pops in more like the main title
