+++ mobileApp/.idea/.name
@@ -1,1 +1,1 @@
-Binge Novel
+Novel binge
