+++ mobileApp/app/src/androidTest/java/com/lebel/novelbinge/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.lebel.bingenovel
+package com.lebel.novelbinge
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.lebel.bingenovel", appContext.packageName)
+        assertEquals("com.lebel.novelbinge", appContext.packageName)
     }
 }
