+++ mobileApp/app/src/main/AndroidManifest.xml
@@ -10,12 +10,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.BingeNovel"
+        android:theme="@style/Theme.NovelBinge"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.BingeNovel">
+            android:theme="@style/Theme.NovelBinge">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
