+++ mobileApp/app/src/main/java/com/lebel/novelbinge/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.lebel.bingenovel
+package com.lebel.novelbinge
 
 import android.os.Bundle
 import com.google.android.material.snackbar.Snackbar
@@ -9,7 +9,7 @@ import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import android.view.Menu
 import android.view.MenuItem
-import com.lebel.bingenovel.databinding.ActivityMainBinding
+import com.lebel.novelbinge.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
@@ -30,8 +30,8 @@ class MainActivity : AppCompatActivity() {
 
         binding.fab.setOnClickListener { view ->
             Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                    .setAction("Action", null)
-                    .setAnchorView(R.id.fab).show()
+                .setAction("Action", null)
+                .setAnchorView(R.id.fab).show()
         }
     }
