+++ mobileApp/app/src/main/java/com/lebel/novelbinge/SecondFragment.kt
@@ -1,4 +1,4 @@
-package com.lebel.bingenovel
+package com.lebel.novelbinge
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
-import com.lebel.bingenovel.databinding.FragmentSecondBinding
+import com.lebel.novelbinge.databinding.FragmentSecondBinding
 
 /**
  * A simple [Fragment] subclass as the second destination in the navigation.
@@ -20,8 +20,8 @@ class SecondFragment : Fragment() {
     private val binding get() = _binding!!
 
     override fun onCreateView(
-            inflater: LayoutInflater, container: ViewGroup?,
-            savedInstanceState: Bundle?
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
     ): View {
 
         _binding = FragmentSecondBinding.inflate(inflater, container, false)
