+++ mobileApp/app/src/main/res/values-v23/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
 
-    <style name="Theme.BingeNovel" parent="Base.Theme.BingeNovel">
+    <style name="Theme.NovelBinge" parent="Base.Theme.NovelBinge">
         <!-- Transparent system bars for edge-to-edge. -->
         <item name="android:navigationBarColor">@android:color/transparent</item>
         <item name="android:statusBarColor">@android:color/transparent</item>
