+++ mobileApp/app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.BingeNovel" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.NovelBinge" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.BingeNovel" parent="Base.Theme.BingeNovel" />
+    <style name="Theme.NovelBinge" parent="Base.Theme.NovelBinge" />
 </resources>
