+++ mobileApp/app/src/test/java/com/lebel/novelbinge/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.lebel.bingenovel
+package com.lebel.novelbinge
 
 import org.junit.Test
